+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `check()` function for checking if text begins with front matter delimiter.
+
+### Changed
+
 ## [0.0.2] - 2018-09-13
 
 ### Changed